requirements.txt
neo4j
python-whois
dnspython
requests
shodan
